+++ src/PipServices.Commons.NetStandard16/Config/MemoryConfigReader.cs
@@ -16,7 +16,7 @@ namespace PipServices.Commons.Config
 
         public virtual Descriptor GetDescriptor()
         {
-            return new Descriptor("pip-services-commons", "config-reader", Name ?? "memory", "1.0");
+            return new Descriptor("pip-services-commons", "config-reader", "memory", Name ?? "default", "1.0");
         }
 
         public virtual void Configure(ConfigParams config)