+++ test/PipServices.Commons.NetStandard16.Test/Refer/DescriptorTest.cs
@@ -8,40 +8,40 @@ namespace PipServices.Commons.Refer
         [Fact]
         public void TestMatch()
         {
-            Descriptor descriptor = new Descriptor("pip-dummies", "controller", "default", "1.0");
+            Descriptor descriptor = new Descriptor("pip-dummies", "controller", "default", "default", "1.0");
 
             // Check match by individual fields
-            Assert.True(descriptor.Match(new Descriptor(null, null, null, null)));
-            Assert.True(descriptor.Match(new Descriptor("pip-dummies", "controller", null, null)));
-            Assert.True(descriptor.Match(new Descriptor(null, null, "default", null)));
-            Assert.True(descriptor.Match(new Descriptor(null, null, null, "1.0")));
+            Assert.True(descriptor.Match(new Descriptor(null, null, null, null, null)));
+            Assert.True(descriptor.Match(new Descriptor("pip-dummies", "controller", null, null, null)));
+            Assert.True(descriptor.Match(new Descriptor(null, null, "default", null, null)));
+            Assert.True(descriptor.Match(new Descriptor(null, null, null, null, "1.0")));
 
             // Check match by individual "*" fields
-            Assert.True(descriptor.Match(new Descriptor("pip-dummies", "*", "*", "*")));
-            Assert.True(descriptor.Match(new Descriptor("*", "controller", "*", "*")));
-            Assert.True(descriptor.Match(new Descriptor("*", "*", "default", "*")));
-            Assert.True(descriptor.Match(new Descriptor("*", "*", "*", "1.0")));
+            Assert.True(descriptor.Match(new Descriptor("pip-dummies", "*", "*", "*", "*")));
+            Assert.True(descriptor.Match(new Descriptor("*", "controller", "*", "*", "*")));
+            Assert.True(descriptor.Match(new Descriptor("*", "*", "default", "*", "*")));
+            Assert.True(descriptor.Match(new Descriptor("*", "*", "*", "*", "1.0")));
 
             // Check match by all values
-            Assert.True(descriptor.Match(new Descriptor("pip-dummies", "controller", "default", null)));
-            Assert.True(descriptor.Match(new Descriptor(null, "controller", "default", "1.0")));
-            Assert.True(descriptor.Match(new Descriptor("pip-dummies", "controller", "default", "1.0")));
+            Assert.True(descriptor.Match(new Descriptor("pip-dummies", "controller", "default", "default", null)));
+            Assert.True(descriptor.Match(new Descriptor(null, "controller", "default", "default", "1.0")));
+            Assert.True(descriptor.Match(new Descriptor("pip-dummies", "controller", "default", "default", "1.0")));
 
             // Check mismatch by individual fields
-            Assert.False(descriptor.Match(new Descriptor(null, "cache", null, null)));
-            Assert.False(descriptor.Match(new Descriptor("pip-services-commons", "controller", null, null)));
-            Assert.False(descriptor.Match(new Descriptor(null, null, "special", null)));
-            Assert.False(descriptor.Match(new Descriptor(null, null, null, "2.0")));
+            Assert.False(descriptor.Match(new Descriptor(null, "cache", null, null, null)));
+            Assert.False(descriptor.Match(new Descriptor("pip-services-commons", "controller", null, null, null)));
+            Assert.False(descriptor.Match(new Descriptor(null, null, "special", null, null)));
+            Assert.False(descriptor.Match(new Descriptor(null, null, null, null, "2.0")));
         }
 
         [Fact]
         public void TestToString()
         {
-            Descriptor descriptor1 = new Descriptor("pip-dummies", "controller", "default", "1.0");
-            Assert.Equal("pip-dummies:controller:default:1.0", descriptor1.ToString());
+            Descriptor descriptor1 = new Descriptor("pip-dummies", "controller", "default", "default", "1.0");
+            Assert.Equal("pip-dummies:controller:default:default:1.0", descriptor1.ToString());
 
-            Descriptor descriptor2 = new Descriptor(null, "controller", null, null);
-            Assert.Equal("*:controller:*:*", descriptor2.ToString());
+            Descriptor descriptor2 = new Descriptor(null, "controller", null, null, null);
+            Assert.Equal("*:controller:*:*:*", descriptor2.ToString());
         }
 
         [Fact]
@@ -49,8 +49,8 @@ namespace PipServices.Commons.Refer
 		    Descriptor descriptor = Descriptor.FromString(null);
 		    Assert.Null(descriptor);
 		
-		    descriptor = Descriptor.FromString("pip-dummies:controller:default:1.0");
-		    Assert.True(descriptor.ExactMatch(new Descriptor("pip-dummies", "controller", "default", "1.0")));
+		    descriptor = Descriptor.FromString("pip-dummies:controller:default:default:1.0");
+		    Assert.True(descriptor.ExactMatch(new Descriptor("pip-dummies", "controller", "default", "default", "1.0")));
 		
 		    try {
 			    Descriptor.FromString("xxx");